+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 ## not released
+- parser: Added optionally distance calculation (fixes #86)
 - parser: Added support for weather data from FANET ground stations
 - parser: Added support for latency in receiver messages (OGNSDR) (fixes #87)
 - parser: Added support for reference_timestamp with tzinfo (fixes #84)